requirements copy.txt
alembic==1.4.3
amqp==5.0.6
apispec==4.4.0
appdirs==1.4.4
arrow==1.0.3
asgiref==3.3.1
asn1crypto==1.4.0
attrs==20.3.0
autopep8==1.5.5
billiard==3.6.4.0
binaryornot==0.4.4
black==20.8b1
blinker==1.4
celery==5.1.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
clipboard==0.0.4
cookiecutter==1.7.2
cycler==0.11.0
distlib==0.3.0
Django==3.1.7
django-extensions==3.1.1
ffmpy==0.3.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Admin==1.5.7
flask-apispec==0.11.0
Flask-Caching==1.9.0
Flask-Cors==3.0.9
Flask-Dance==3.2.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
git+https://github.com/ashcrow/flask-track-usage@629ad1c#egg=Flask-Track-Usage
fonttools==4.29.0
GeoAlchemy2==0.8.4
greenlet==1.0.0
idna==2.10
importlib-metadata==4.8.1
importlib-resources==5.2.3
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jinja2-time==0.2.0
kiwisolver==1.3.2
kombu==5.1.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-polyfield==5.10
matplotlib==3.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.4
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
pg8000==1.19.2
Pillow==8.0.1
pipenv==2020.11.15
pluggy==0.13.1
poyo==0.5.0
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pygeohash==1.2.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pytest==6.2.3
pytest-flask==1.2.0
pytest-timeout==1.4.2
python-dateutil==2.8.1
python-editor==1.0.4
python-http-client==3.3.2
python-magic==0.4.24
python-slugify==4.0.1
PythonVideoConverter==1.0.3
pytz==2020.4
regex==2021.4.4
requests==2.25.0
requests-oauthlib==1.3.0
reverse-geocode==1.4.1
scipy==1.5.4
scramp==1.4.0
sendgrid==6.6.0
sentry-sdk==0.19.4
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Continuum==1.3.11
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
testing.common.database==2.0.3
testing.postgresql==1.3.0
text-unidecode==1.3
timezonefinder==4.5.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
URLObject==2.4.3
vine==5.0.0
virtualenv==20.0.23
virtualenv-clone==0.5.4
wcwidth==0.2.5
webargs==8.0.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.6.0
